Small fixes in examples.

+++ examples/layout_hyper_mesh.c
@@ -174,6 +174,6 @@ get_coords(int x[], int node, int index)
 
 
 /* The number of sites on the specified node */
-size_t num_sites(int node){
+int num_sites(int node){
   return sites_on_node;
 }
