+++ examples/qio-convert-mesh.h
@@ -8,7 +8,7 @@ int setup_layout(int len[], int nd, int nsquares2[], int ndim2);
 int node_number(const int x[]);
 int node_index(const int x[]);
 void get_coords(int x[], int node, int index);
-size_t num_sites(int node);
+int num_sites(int node);
 
 typedef struct
 {
